+++ node/impl/full/eth.go
@@ -177,6 +177,16 @@ type EthAPI struct {
 	EthEventAPI
 }
 
+type EthCallError struct {
+	Message string
+	Code    int
+	Data    string
+}
+
+func (e *EthCallError) Error() string {
+	return e.Data
+}
+
 var ErrNullRound = errors.New("requested epoch was a null round")
 
 func (a *EthModule) StateNetworkName(ctx context.Context) (dtypes.NetworkName, error) {
@@ -1433,9 +1443,9 @@ func (a *EthModule) applyMessage(ctx context.Context, msg *types.Message, tsk ty
 	}
 
 	if res.MsgRct.ExitCode.IsError() {
-		reason := parseEthRevert(res.MsgRct.Return)
-		return nil, xerrors.Errorf("message execution failed: exit %s, revert reason: %s, vm error: %s", res.MsgRct.ExitCode, reason, res.Error)
+		return nil, errors.New(parseEthRevert(res.MsgRct.Return))
 	}
+
 	return res, nil
 }
 
@@ -1473,10 +1483,15 @@ func (a *EthModule) EthEstimateGas(ctx context.Context, p jsonrpc.RawParams) (et
 		// guts of EthCall). This will give us an ethereum specific error with revert
 		// information.
 		msg.GasLimit = buildconstants.BlockGasLimit
-		if _, err2 := a.applyMessage(ctx, msg, ts.Key()); err2 != nil {
+		invokeResult, err2 := a.applyMessage(ctx, msg, ts.Key())
+		if err2 != nil {
 			err = err2
 		}
-		return ethtypes.EthUint64(0), xerrors.Errorf("failed to estimate gas: %w", err)
+		return ethtypes.EthUint64(0), &EthCallError{
+			Message: fmt.Sprintf("failed to estimate gas: %s", err),
+			Code:    -32000,
+			Data:    parseEthRevert(invokeResult.MsgRct.Return),
+		}
 	}
 
 	expectedGas, err := ethGasSearch(ctx, a.Chain, a.Stmgr, a.Mpool, gassedMsg, ts)
@@ -1614,24 +1629,58 @@ func ethGasSearch(
 func (a *EthModule) EthCall(ctx context.Context, tx ethtypes.EthCall, blkParam ethtypes.EthBlockNumberOrHash) (ethtypes.EthBytes, error) {
 	msg, err := ethCallToFilecoinMessage(ctx, tx)
 	if err != nil {
-		return nil, xerrors.Errorf("failed to convert ethcall to filecoin message: %w", err)
+		return nil, ethtypes.EthCallErrorWithDefaultCode(
+			fmt.Sprintf("failed to convert ethcall to filecoin message"),
+			err.Error(),
+		)
 	}
 
 	ts, err := getTipsetByEthBlockNumberOrHash(ctx, a.Chain, blkParam)
 	if err != nil {
-		return nil, xerrors.Errorf("failed to process block param: %v; %w", blkParam, err)
+		return nil, ethtypes.EthCallErrorWithDefaultCode(
+			fmt.Sprintf("failed to process block param: %v; %s", blkParam),
+			err.Error(),
+		)
 	}
 
 	invokeResult, err := a.applyMessage(ctx, msg, ts.Key())
 	if err != nil {
-		return nil, err
+		return nil, ethtypes.EthCallErrorWithDefaultCode(
+			fmt.Sprintf("execution revereted"),
+			err.Error(),
+		)
+	}
+
+	if invokeResult == nil {
+		return nil, ethtypes.EthCallErrorWithDefaultCode(
+			fmt.Sprintf("execution revereted"),
+			"invoke result is nil",
+		)
+	}
+
+	if invokeResult.MsgRct == nil {
+		return ethtypes.EthBytes{}, nil
 	}
 
 	if msg.To == builtintypes.EthereumAddressManagerActorAddr {
-		// As far as I can tell, the Eth API always returns empty on contract deployment
 		return ethtypes.EthBytes{}, nil
 	} else if len(invokeResult.MsgRct.Return) > 0 {
-		return cbg.ReadByteArray(bytes.NewReader(invokeResult.MsgRct.Return), uint64(len(invokeResult.MsgRct.Return)))
+		result := ethtypes.EthBytes{}
+		defer func() {
+			if r := recover(); r != nil {
+				result = ethtypes.EthBytes{}
+			}
+		}()
+
+		reader := bytes.NewReader(invokeResult.MsgRct.Return)
+		result, err = cbg.ReadByteArray(reader, uint64(len(invokeResult.MsgRct.Return)))
+		if err != nil {
+			return nil, ethtypes.EthCallErrorWithDefaultCode(
+				fmt.Sprintf("execution revereted"),
+				err.Error(),
+			)
+		}
+		return result, nil
 	}
 
 	return ethtypes.EthBytes{}, nil